code/python/py-nats-dlq-demo/requirement.txt
nats-py>=2.2.0
